+++ CHANGELOG.md
@@ -43,3 +43,7 @@
 
 - Updated the getFees() method for gas estimation and added its test.
 - Changed the License to MIT.
+
+### 1.0.9 (2023-12-11)
+
+- Added gas estimation for EIP-1559 transaction on Optimism.